feat: Initial operations logic (#298)
Adds the initial scaffolding for running operations. It's structured in
a similar way to the original, but with lots of stuff trimmed out.

I've implemented two example operations, `DISTINCT` which returns a SQL
query which is a list and `MAX` which is a SQL query returning a single
value

The actual `output_variables` are passed through to the class where the
check operators are executed but not used yet. Will come in another PR