chore: Refactor the DBCache (#301)
We now track postgres schema in-memory. ie we store the name, hash and
type of each column, grouped into tables, grouped into the whole db.

The schema is stored in the postgres interface itself, so there's no
longer the separation between making a new column in the cache and
adding the column.

This is useful because we may need the column type for certain
operations (ie `empty`, where the query depends on the column type)