chore: domain-based rule skipping (#305)
A while ago, when we were refactoring during the end-step of PostgresQL
Data Service development, a large portion of the sql_rule_processor.py
was commented out. This removed some of the skipping functionality that
detects and verifies rules which either shouldn't be run on certain
data, or doesn't need to be run at all. The reason this functionality
existed beforehand was to effectively minimise the run time of running
all rules on all data (e.g. regression tests, real study sweep tests).

This branch and associated PRs aims to bring that functionality back. 

This PR is dedicated to uncommenting and also restoring the main
skipping framework and within that, the domain-based skipping logic. The
logic for class-based, use-case-based and rule-pattern-based skipping
will be implemented in future PRs on this branch.

The edits to the regression file are being fixed as a List/dict type for
result is wrong.

---------

Co-authored-by: YetiGonk <maximo@verisian.com>