chore: class-based rule skipping (#322)
A while ago, when we were refactoring during the end-step of PostgresQL
Data Service development, a large portion of the sql_rule_processor.py
was commented out. This removed some of the skipping functionality that
detects and verifies rules which either shouldn't be run on certain
data, or doesn't need to be run at all. The reason this functionality
existed beforehand was to effectively minimise the run time of running
all rules on all data (e.g. regression tests, real study sweep tests).

This branch and associated PRs aim to bring that functionality back.

This PR is dedicated to uncommenting and also restoring the class-based
skipping logic. The logic for use-case-based skipping will be
implemented in the final PR on this branch.

---------

Co-authored-by: YetiGonk <maximo@verisian.com>