chore: RULE PROCESSOR: use-case based skipping logic (#340)
A while ago, when we were refactoring during the end-step of PostgresQL
Data Service development, a large portion of the sql_rule_processor.py
was commented out. This removed some of the skipping functionality that
detects and verifies rules which either shouldn't be run on certain
data, or doesn't need to be run at all. The reason this functionality
existed beforehand was to effectively minimise the run time of running
all rules on all data (e.g. regression tests, real study sweep tests).

This branch and associated PRs aim to bring that functionality back.

This PR is dedicated to use-case based skipping logic. This is the final
LOGIC PR on this branch.

I have removed use-case skipping logic because it is TIG specific. We
are working with SDTMIG and ADaM rules, not TIG. This was an old feature
in CORE even though they never had TIG. Perhaps we reimplement once we
have a complete CORE but for now, it serves no purpose to us.

---------

Co-authored-by: YetiGonk <maximo@verisian.com>