fix: Improve joins (#345)
Makes it so that when two tables are joined, all of the columns in the
second table are retained but stored as `<right table name>.column`. If
the column name is not present in the left table, then it also creates
an alias for that column under the same name. So a rule could reference
either `ABC` or `R.ABC` and both work