fix: More date operator fixes (#344)
This PR introduces a new method which takes a string column and
generates a timestamp column by casting the string column. Any invalid
dates are set to null.

It then updates the `is_complete_date` and `date_comparison` operators
to use that generated column. This fixes a problem where the date
comparison operators would try to cast columns containing invalid dates
and throw a postgres error