fix: Dataset metadata input (#405)
Standardises the input format for dataset metadata. Passes those changes
through to the data service but doesn't replace the SDTMMetadata object
used by the rule editor, that will be done in another PR as this is
already big enough.

This PR breaks a number of rules, primarily the rules which rely on
relationship merges, the logic for relationship merges needs to be moved
into the SDTM standard class