fix: study domain sort in regression (#412)
Study domain lists in the regression rules.json frequently change their
order from PR to PR. While easy to spot and diagnose as not related to
whatever PR they change on, it requires a clean-up. That is the purpose
of this PR.

I have added a small sort to the old and sql errors before they are
appended to the larger regression dict to maintain their order.

I have also moved the error translation from #409 to a separate helper
function to skirt around prettier's complexity flag. This also makes it
easier for us to remove error translation later down the line when it is
no longer needed.

(((regression has been run for testing but has not been committed
because of looming merge conflicts with other regression PRs currently
up)))

---------

Co-authored-by: YetiGonk <maximo@verisian.com>