feat: Rule Regression Whitelist (#411)
Whitelisted rules are rules where the fault and therefore diff lies in
the old engine's code. It is a rule where our new sql engine's result is
correct but a diff between new and old is still present. We do not
intend to fix old engine code to match our new engine and so this is the
resulting solution. This is also for the purpose of weeding out the rule
diffs that have already been reviewed and addressed as well as making
the regression pie chart cleaner and closer to a match between engines.

A follow-up PR to add a "whitelisted" sensitivity to the pie chart
displays should be created.

Co-authored-by: YetiGonk <maximo@verisian.com>