feat: OPERATION: extract_metadata (#419)
Added extract_metadata operation and tests

Created metadata_mappings dictionary to map attribute names - the rules
use 'dataset_name' but we use 'name' attribute on dataset objects. I
suspect this problem would recur if extract_metadata was more widely
used, so the dictionary seemed a good solution.

Ran regression

Note that many regression issues are due to to a rule error where
'value_is_literal: true" has been missed, so it's trying to find the
dataset as a column. We have it noted to fix the rule shortly.