fix: get cmd working with -sql (#426)
Small tweaks to get the -sql flag working. It now produces an identical
output report as without -sql flag (with one exception - we don't store
many dataset attributes in our metadata, so they are populating as blank
in the 'Dataset Details' tab), but it runs using our engine and thus
more quickly

Very small and simple changes - excel_report adjustments are mainly just
formatting

UPDATE:
Got all but one QARegressionTests working for -sql. Some are probably
not needed for our use case but figured better to get as many as
possible working and then remove those not needed.

The one that's commented out is a log-level checker, and I think the
alias issue from _column_sql is causing issues there (raising errors
that the test thinks shouldn't be present).