feat: Optional Variables (#427)
This PR aims to implement an additional check in base sql operator that
variables being accessed by a check operator exist. When a check
operator looks for a variable that doesn’t exist, it throws an error.
This obviously invalidates the rule. Now, if a column/variable doesn't
exist, they are handled uniquely by the check operator to return either
a sql TRUE or FALSE depending on the context.

In certain rules, we sometimes have multiple check operators running one
after another. These can be wrapped in an any or all like such:

```
Check:
  all:
    - name: dataset_name
      operator: matches_regex
      value: ^[a-z(?-i)]{2}[a-z0-9(?-i)]{1,2}
    - name: dataset_name
      operator: not_prefix_matches_regex
      prefix: 2
      value: (AP|FA)
    - name: dataset_name
      operator: prefix_is_not_contained_by
      prefix: 2
      value: $list_dataset_names
```

For example, `is_not_complete_date` on a non-existent column should be
marked as `TRUE` because it’s not a complete date, the column where the
date would be doesn’t exist.

All of the regression changes line up with these operator changes.

---------

Co-authored-by: YetiGonk <maximo@verisian.com>