feat: Add Table Prefix CLI arg (#430)
Adds the CLI argument which adds a unique `sql_namespace` to all table
hashes when running the main workflow.
This sql_namespace is appended to the start of table hashes that are
added to the database.
All previous non-static tables that have this prefix are dropped from
the database upon initialisation.
The default for an unspecified table_prefix is an 8 char uid that is
Postgres compatible.

Regression has not changed at all (to be expected, we do not add the
internal table_name to any part of the regression json).
Unit test change to amend use of table_id in sql query instead of hash.

---------

Co-authored-by: YetiGonk <maximo@verisian.com>